fractal-navigation: update to remove fractal-navigation-router components
