+++ packages/fractal-navigation/src/TabBar/components/CircularTabBarItem/CircularTabBarItemLink.tsx
@@ -1,18 +1,14 @@
 import React from 'react';
-import { useIsRouteActive, useGoToTab } from '@bma98/fractal-navigation-router';
 import { BottomCircularTabBarItem } from './BottomCircularTabBarItem';
 import { LeftCircularTabBarItem } from './LeftCircularTabBarItem';
 import { RightCircularTabBarItem } from './RightCircularTabBarItem';
 
 interface CircularTabBarItemLinkProps {
-    path: string;
     tabBarPosition: 'left' | 'bottom' | 'right';
+    goToTab: () => void;
 }
 
-export function CircularTabBarItemLink({ tabBarPosition, path }: CircularTabBarItemLinkProps): JSX.Element {
-    const isRouteActive = useIsRouteActive(path, false);
-    const goToTab = useGoToTab(path, isRouteActive);
-
+export function CircularTabBarItemLink({ tabBarPosition, goToTab }: CircularTabBarItemLinkProps): JSX.Element {
     switch (tabBarPosition) {
         case 'left':
             return <LeftCircularTabBarItem onPress={goToTab} />;
